+++ src/backend/src/public/gamification-api-docs.html
@@ -3,16 +3,12 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <title>Clamp API Docs</title>
+    <title>Clamp Gamification API Docs</title>
     <!-- Embed elements Elements via Web Component -->
     <script src="https://unpkg.com/@stoplight/elements/web-components.min.js"></script>
     <link rel="stylesheet" href="https://unpkg.com/@stoplight/elements/styles.min.css" />
   </head>
   <body>
-    <elements-api
-      apiDescriptionUrl="/insights-api/spec"
-      router="hash"
-      layout="sidebar"
-    />
+    <elements-api apiDescriptionUrl="/gamification-api/spec" router="hash" layout="sidebar" />
   </body>
 </html>
