+++ src/pages/gamifiedCampaign/new.tsx
@@ -22,7 +22,7 @@ import { createNewCampaign } from '@/httpClient/campaign'
 import { createNewRule } from '@/httpClient/rules'
 import { initNewGame } from '@/httpClient/game'
 import { IDraftGame } from '@/backend/src/v1/gamificationAPI/interfaces/IGame'
-import { winningCriteria } from '@/backend/src/v1/gamificationAPI/services/game'
+import { winningCriteria } from '@/backend/src/lib/game'
 interface NewGamifiedCampaignFormValues {
   campaignName: string
   roundsDuration: string
